Improved Robot network table communication
Converted RobotCodeCommununication to an AppState, added a loop to
update pairings between the network table and the sim code so that
variables will take on values from the network table (robot code) and
update values to the network table (for sensors that need to pass back
information). Also added an unpair command to delink sim objects and
network table data.