RobotCodeCommunication now extends thread (untested), encoders improved
RobotCodeCommunication is again a singleton, extending thread instead of
AppState in order to improve the speed of our update loops and possibly
reduce latency. The improvements on the encoders is due to the encoders
now taking into account which side of the robot they are on, so a right
turn will properly move the left encoder more than the right encoder
instead of both sides having equal encoder readings.