Connection to network table properly closes on exit now, fixed jar
Jar Fix - SendableChooser will now go onto network tables properly after
building robot code with fake libraries, goes on as subtable now instead
of a series of independent keys. Added an override in SimMain of legacy
Application and some corresponding methods in RobotCodeCommunication to
close connection if it was running.