Changing some setup things, allows Sim to be packaged as jar
The program is now run out of a different class so that the libraries
can be loaded in without the states depending of ntcore getting
initialized first. The data folder was moved into the assets folder so
that it can appear inside the jar and be copied out into a separate
folder for setup and running with native libraries.